Refactor packages for neat setup/teardown
