+++ ml-data-hub-plugin/src/test/groovy/com/marklogic/gradle/fullcycle/NotInstalledTests.groovy
@@ -15,22 +15,23 @@
  *
  */
 
-package com.marklogic.gradle.task
+package com.marklogic.gradle.fullcycle
 
+import com.marklogic.gradle.task.BaseTest
 import org.gradle.testkit.runner.UnexpectedBuildSuccess
 
 import static org.gradle.testkit.runner.TaskOutcome.FAILED
 
-class ButThisIsNotInstalledTests extends BaseTest {
+class NotInstalledTests extends BaseTest {
 
     def setupSpec() {
         createGradleFiles()
         runTask('hubInit')
-        runTask('mlUndeploy',  '-Pconfirm=true')
+        //runTask('mlUndeploy',  '-Pconfirm=true')
     }
 
     def cleanupSpec() {
-        runTask('mlDeploy')
+        //runTask('mlDeploy')
     }
 
     def "enable debugging hub not installed"() {
@@ -65,7 +66,7 @@ class ButThisIsNotInstalledTests extends BaseTest {
 
     def "test run flow with no flow"() {
         given:
-        propertiesFile << """
+        BaseTest.propertiesFile << """
             ext {
                 entityName=my-entity
             }
@@ -82,7 +83,7 @@ class ButThisIsNotInstalledTests extends BaseTest {
 
     def "test run flow when hub not installed"() {
         given:
-        propertiesFile << """
+        BaseTest.propertiesFile << """
                 ext {
                     entityName=my-entity
                     flowName=my-flow
