+++ ml-data-hub-plugin/src/test/groovy/com/marklogic/gradle/fullcycle/SslTest.groovy
@@ -15,12 +15,13 @@
  *
  */
 
-package com.marklogic.gradle.task
+package com.marklogic.gradle.fullcycle
 
 import com.marklogic.client.DatabaseClientFactory
 import com.marklogic.client.ext.modulesloader.ssl.SimpleX509TrustManager
 import com.marklogic.client.io.DOMHandle
 import com.marklogic.client.io.DocumentMetadataHandle
+import com.marklogic.gradle.task.BaseTest
 import com.marklogic.hub.HubConfig
 import org.gradle.testkit.runner.UnexpectedBuildFailure
 
@@ -38,8 +39,8 @@ import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 class SslTest extends BaseTest {
     def setupSpec() {
         createFullPropertiesFile()
-        buildFile = testProjectDir.newFile('build.gradle')
-        buildFile << '''
+        BaseTest.buildFile = BaseTest.testProjectDir.newFile('build.gradle')
+        BaseTest.buildFile << '''
             plugins {
                 id 'com.marklogic.ml-data-hub'
             }
@@ -131,14 +132,14 @@ class SslTest extends BaseTest {
         '''
 
         runTask("hubInit")
-        copyResourceToFile("ssl-test/my-template.xml", new File(testProjectDir.root, "user-config/security/certificate-templates/my-template.xml"))
-        copyResourceToFile("ssl-test/ssl-server.json", new File(testProjectDir.root, "user-config/servers/final-server.json"))
-        copyResourceToFile("ssl-test/ssl-server.json", new File(testProjectDir.root, "user-config/servers/job-server.json"))
-        copyResourceToFile("ssl-test/ssl-server.json", new File(testProjectDir.root, "user-config/servers/staging-server.json"))
-        copyResourceToFile("ssl-test/ssl-server.json", new File(testProjectDir.root, "user-config/servers/trace-server.json"))
+        copyResourceToFile("ssl-test/my-template.xml", new File(BaseTest.testProjectDir.root, "user-config/security/certificate-templates/my-template.xml"))
+        copyResourceToFile("ssl-test/ssl-server.json", new File(BaseTest.testProjectDir.root, "user-config/servers/final-server.json"))
+        copyResourceToFile("ssl-test/ssl-server.json", new File(BaseTest.testProjectDir.root, "user-config/servers/job-server.json"))
+        copyResourceToFile("ssl-test/ssl-server.json", new File(BaseTest.testProjectDir.root, "user-config/servers/staging-server.json"))
+        copyResourceToFile("ssl-test/ssl-server.json", new File(BaseTest.testProjectDir.root, "user-config/servers/trace-server.json"))
         createProperties()
         try {
-            clearDatabases("data-hub-MODULES")
+            clearDatabases(hubConfig().DEFAULT_MODULES_DB_NAME)
         } catch (e) {
             //pass
         }
@@ -152,8 +153,8 @@ class SslTest extends BaseTest {
 
 
     void createProperties() {
-        propertiesFile = new File(testProjectDir.root, 'gradle.properties')
-        propertiesFile << """
+        BaseTest.propertiesFile = new File(BaseTest.testProjectDir.root, 'gradle.properties')
+        BaseTest.propertiesFile << """
         mlAdminScheme=https
         mlManageScheme=https
         mlAdminSimpleSsl=true
@@ -179,7 +180,7 @@ class SslTest extends BaseTest {
         then:
         notThrown(UnexpectedBuildFailure)
         def modCount = getModulesDocCount()
-        modCount == MOD_COUNT_WITH_TRACE_MODULES || modCount == MOD_COUNT
+        modCount == BaseTest.MOD_COUNT_WITH_TRACE_MODULES || modCount == BaseTest.MOD_COUNT
         result.task(":mlDeploy").outcome == SUCCESS
     }
