+++ ml-data-hub-plugin/src/test/groovy/com/marklogic/gradle/task/InstalledTests.groovy
@@ -35,8 +35,8 @@ class InstalledTests extends BaseTest {
     def setupSpec() {
         createGradleFiles()
         runTask('hubInit')
-        //runTask('mlUndeploy',  '-Pconfirm=true')
-        //println(runTask('mlDeploy', '-i').getOutput())
+        // this will be relatively fast (idempotent) for already-installed hubs
+        println(runTask('mlDeploy', '-i').getOutput())
     }
 
     def cleanupSpec() {
